gembah1-invoicing/requirements.txt
rpaframework==10.*
robotframework-browser
ta-bitwarden-cli
slackclient
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
datetime
retry
